pyproject.toml
[project]
name = "unit-testing-sql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.1",
]

[tool.uv]
package = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["app*"]

[tool.setuptools.package-data]
app = ["sql/*.sql"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
